+++ crates/teloxide/src/dispatching/dialogue/storage/postgres_storage/queries/keys.sql
@@ -1,0 +1,1 @@
+SELECT chat_id FROM teloxide_dialogues