+++ crates/teloxide/src/dispatching/dialogue/storage/sqlite_storage.rs
@@ -3,6 +3,7 @@ use futures::future::BoxFuture;
 use serde::{de::DeserializeOwned, Serialize};
 use sqlx::{sqlite::SqlitePool, Executor};
 use std::{
+    collections::HashSet,
     convert::Infallible,
     fmt::{Debug, Display},
     str,
@@ -119,6 +120,24 @@ where
                 .await?
                 .map(|d| self.serializer.deserialize(&d).map_err(SqliteStorageError::SerdeError))
                 .transpose()
+        })
+    }
+
+    fn keys(self: Arc<Self>) -> BoxFuture<'static, Result<HashSet<ChatId>, Self::Error>> {
+        Box::pin(async move {
+            #[derive(sqlx::FromRow)]
+            struct DialogueDbRow {
+                chat_id: i64,
+            }
+
+            let keys = sqlx::query_as::<_, DialogueDbRow>("SELECT chat_id FROM teloxide_dialogues")
+                .fetch_all(&self.pool)
+                .await?
+                .iter()
+                .map(|r| ChatId(r.chat_id))
+                .collect();
+
+            Ok(keys)
         })
     }
 }